12: index into gmem blocks, remove advance block, normalize O by multiplying reciprocal of l

+++ src/include/array.cuh
@@ -18,10 +18,12 @@ struct Array {
 
     FA_DEVICE_CONSTEXPR void zero() { fill(0); }
 
-    FA_DEVICE_CONSTEXPR value_t &operator[](int idx) { return _data[idx]; }
-    FA_DEVICE_CONSTEXPR value_t operator[](int idx) const { return _data[idx]; }
+    FA_DEVICE_CONSTEXPR value_t &operator[](size_t idx) { return _data[idx]; }
+    FA_DEVICE_CONSTEXPR value_t operator[](size_t idx) const {
+        return _data[idx];
+    }
 
-    FA_DEVICE_CONSTEXPR static int size() { return N; }
+    FA_DEVICE_CONSTEXPR static size_t size() { return N; }
 };
 
 template <int N, typename value_t, int Alignment = 16>
